moved login and login css out of components. updated css, and updated font globally.
