+++ client/main.css
@@ -1,4 +1,4 @@
 body {
   padding: 10px;
-  font-family: sans-serif;
+  font-family: "Roboto Condensed", sans-serif;
 }
