+++ client/main.html
@@ -1,5 +1,9 @@
 <head>
   <title>digital-bulletin-admin-tool</title>
+  <link
+    href="https://fonts.googleapis.com/css2?family=Roboto+Condensed:wght@300&display=swap"
+    rel="stylesheet"
+  />
 </head>
 
 <body>
